updated api list for user notion

+++ apps/server/src/components/list/list-controllers.js
@@ -4,7 +4,8 @@ import { SchemaType } from 'mongoose'
 
 export const getLists = async (ctx) => {
   try {
-    const lists = await ListModel.find({})
+    const userId = ctx.state.user._id
+    const lists = await ListModel.find({ user: userId })
     ctx.ok(lists)
   } catch (error) {
     ctx.badRequest({ message: error.message })
@@ -13,8 +14,9 @@ export const getLists = async (ctx) => {
 
 export const getOneList = async (ctx) => {
   try {
-    const list = await ListModel.findOne({ _id: ctx.params.id })
-    if(!list) throw new Error('List not found')
+    const user = ctx.state.user;
+    const list = await ListModel.findOne({ _id: ctx.params.id, user: user._id });
+    if (!list) throw new Error('List not found')
     ctx.ok(list)
   } catch (error) {
     ctx.badRequest({ message: error.message })
@@ -32,10 +34,11 @@ export const createList = async (ctx) => {
 
     if(error) throw new Error(error)
 
-    const listname = await ListModel.findOne({ title: ctx.request.body.title })
+    const { userId } = ctx.state.user
+    const listname = await ListModel.findOne({ title: ctx.request.body.title, userId })
     if (listname) throw new Error('List already exists')
 
-    const list = await ListModel.create(value)
+    const list = await ListModel.create({ ...value, userId })
     ctx.ok(list)
   } catch (error) {
     ctx.badRequest({ message: error.message })
@@ -44,6 +47,7 @@ export const createList = async (ctx) => {
 
 export const updateList = async (ctx) => {
   try {
+    const userId = ctx.state.user.id;
     if (!ctx.request.body.title) throw new Error('Title is missing')
 
     const schema = joi.object({
@@ -53,11 +57,13 @@ export const updateList = async (ctx) => {
 
     if(error) throw new Error(error)
 
-    const listname = await ListModel.findOne({ title: ctx.request.body.title })
-    if (listname) throw new Error('List already exists')
+    const list = await ListModel.findOne({ _id: ctx.params.id, user: userId });
+    if (!list) throw new Error('List not found');
+    const listname = await ListModel.findOne({ title: ctx.request.body.title, user: userId });
+    if (listname && listname._id.toString() !== list._id.toString()) throw new Error('List already exists');
   
-    const list = await ListModel.findOneAndUpdate({ _id: ctx.params.id }, value, { new: true })
-    ctx.ok(list)
+    const updatedList = await ListModel.findOneAndUpdate({ _id: ctx.params.id, user: userId }, value, { new: true });
+    ctx.ok(updatedList)
   } catch (error) {
     ctx.badRequest({ message: error.message })
   }
@@ -65,8 +71,10 @@ export const updateList = async (ctx) => {
 
 export const deleteList = async (ctx) => {
   try {
-    const list = await ListModel.findOneAndDelete({ _id: ctx.params.id })
-    if(!list) throw new Error('List not found')
+    const userId = ctx.state.user._id
+
+    const list = await ListModel.findOneAndDelete({ _id: ctx.params.id, user: userId })
+    if(!list) throw new Error('List not found or user is not authorized to delete it')
     ctx.ok(list)
   } catch (error) {
     ctx.badRequest({ message: error.message })
@@ -75,17 +83,19 @@ export const deleteList = async (ctx) => {
 
 export const getTasksByListId = async (ctx) => {
   try {
-    const tasks = await TaskModel.find({ list: ctx.params.id })
-    if(!tasks) throw new Error('List not found')
-    ctx.ok(tasks)
+    const list = await ListModel.findOne({ _id: ctx.params.id, user: ctx.state.user._id });
+    if (!list) throw new Error('List not found');
+
+    const tasks = await TaskModel.find({ list: list._id });
+    ctx.ok(tasks);
   } catch (error) {
-    ctx.badRequest({ message: error.message })
+    ctx.badRequest({ message: error.message });
   }
-}
+};
 
 export const getCompletedTasksByListId = async (ctx) => {
   try {
-    const tasks = await TaskModel.find({ list: ctx.params.id, state: true })
+    const tasks = await TaskModel.find({ list: ctx.params.id, state: true, user: ctx.state.user.id })
     ctx.ok(tasks || [])
   } catch (error) {
     ctx.badRequest({ message: error.message })
@@ -94,7 +104,7 @@ export const getCompletedTasksByListId = async (ctx) => {
 
 export const getUncompletedTasksByListId = async (ctx) => {
   try {
-    const tasks = await TaskModel.find({ list: ctx.params.id, state: false })
+    const tasks = await TaskModel.find({ list: ctx.params.id, state: false, user: ctx.state.user.id })
     ctx.ok(tasks || [])
   } catch (error) {
     ctx.badRequest({ message: error.message })
